grim_analysis+grim_engine+grim_telemetry_common: surface tag and handle labels in retail telemetry
Why:
- ease composite inference from raw logs with tag/handle names
- capture mutator callers to locate table/global writes
- keep rust engine schema aligned with shim

What:
- grim_telemetry_common/src/lib.rs: add tag_label/handle_label/caller fields and tests
- grim_analysis/retail_capture/rust_shim/src/trace.rs: track tag labels from settagmethod, cache handle labels, emit caller on mutators
- grim_analysis/retail_capture/rust_shim/README.md: document tag_label and handle_label coverage plus caller fields
- grim_engine/src/lua_host/telemetry.rs: forward new optional fields in engine emitter
- grim_engine/src/lua_host/context/bindings/bootstrap.rs: include handle_label when logging type global
- grim_analysis/retail_capture/rust_shim/src/lib.rs: ensure pushnumber export uses f32 signature
- grim_analysis/retail_capture/rust_shim/src/lua_api.rs: align pushnumber FFI to f32

+++ grim_analysis/retail_capture/rust_shim/src/lib.rs
@@ -87,7 +87,7 @@ pub unsafe extern "C" fn lua_error(message: *const libc::c_char) {
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn lua_pushnumber(value: libc::c_double) {
+pub unsafe extern "C" fn lua_pushnumber(value: libc::c_float) {
     trace_lua_pushnumber(value);
 }
